add support for g2.large.x86
This is almost a 1:1 copy of the x2.xlarge.x86 expression. The `maxJobs`
attribute has been adjusted to fit the hardware.

The hardware is mostly identical to the x2.xlarge.x86 instance for our
purposes. It comes with one SSD that we use for the OS (as long as it
comes first aka is sda). The larger ones disks are not touched during
the installation - as usual.